+++ DrillingRig.Commands/AinSettings/IAinSettings.cs
@@ -6,20 +6,20 @@ namespace DrillingRig.Commands.AinSettings {
 		BytesPair Reserved00 { get; }
 
 		decimal KpW { get; }
-		double KiW { get; }
+		decimal KiW { get; }
 
-		double FiNom { get; }
+		decimal FiNom { get; }
 		short Imax { get; }
 		short UdcMax { get; }
 		short UdcMin { get; }
-		double Fnom { get; }
-		double Fmax { get; }
+		decimal Fnom { get; }
+		decimal Fmax { get; }
 
-		double DflLim { get; }
-		double FlMinMin { get; }
+		decimal DflLim { get; }
+		decimal FlMinMin { get; }
 		
 		short IoutMax { get; }
-		double FiMin { get; }
+		decimal FiMin { get; }
 
 		ushort DacCh { get; }
 		ushort Imcw { get; }
@@ -30,42 +30,42 @@ namespace DrillingRig.Commands.AinSettings {
 
 		short Udc0 { get; }
 
-		double TauR { get; }
-		double Lm { get; }
-		double Lsl { get; }
-		double Lrl { get; }
+		decimal TauR { get; }
+		decimal Lm { get; }
+		decimal Lsl { get; }
+		decimal Lrl { get; }
 
 		BytesPair Reserved24 { get; }
-		double KpFi { get; }
-		double KiFi { get; }
+		decimal KpFi { get; }
+		decimal KiFi { get; }
 
 		BytesPair Reserved28 { get; }
-		double KpId { get; }
-		double KiId { get; }
+		decimal KpId { get; }
+		decimal KiId { get; }
 
 		BytesPair Reserved32 { get; }
-		double KpIq { get; }
-		double KiIq { get; }
+		decimal KpIq { get; }
+		decimal KiIq { get; }
 
-		double AccDfDt { get; }
-		double DecDfDt { get; }
+		decimal AccDfDt { get; }
+		decimal DecDfDt { get; }
 
 		//short Unom { get; }
-		double Unom { get; }
+		decimal Unom { get; }
 
 		/// <summary>
 		/// Постоянная времени регулятора компенсации потока
 		/// </summary>
-		double TauFlLim { get; }
+		decimal TauFlLim { get; }
 
-		double Rs { get; }
+		decimal Rs { get; }
 
-		double Fmin { get; }
+		decimal Fmin { get; }
 
-		double TauM { get; }
-		double TauF { get; }
-		double TauFSet { get; }
-		double TauFi { get; }
+		decimal TauM { get; }
+		decimal TauF { get; }
+		decimal TauFSet { get; }
+		decimal TauFi { get; }
 
 		short IdSetMin { get; }
 		short IdSetMax { get; }
@@ -87,14 +87,14 @@ namespace DrillingRig.Commands.AinSettings {
 		//
 		int NimpFloorCode { get; }
 		AinTelemetryFanWorkmode FanMode { get; }
-		
+
 
 		/// <summary>
 		/// Порог компенсации напряжения DC за счет потока
 		/// </summary>
-		double UmodThr { get; }
+		decimal UmodThr { get; }
 
-		double EmdecDfdt { get; }
+		decimal EmdecDfdt { get; }
 		short TextMax { get; }
 		short ToHl { get; }
